[BASE] pkmnsv_ev_tracker - WIP
This project has one goal: make the EV training easier to follow in
Pokémon Scarlet & Violet.

For now, it's a WIP, so this commit will be squashed a lot of time, at
least until I have someting usable.

The planned features are:
[x] Track up to 6 Pokémons EV at the same time;
[ ] Chose your Pokémon specy;
[ ] Rename your Pokémon;
[ ] Save progress in the localStorage to get back your Pokémons when the
page is opened;
[ ] Click on defaited Pokémon to gain their corresponding EV;
  [ ] Filter Pokémon by game-version;
  [ ] Filter Pokémon by DLC/base game only;
  [ ] Filter Pokémon by given EV;
[ ] Upload this project somewhere to make it usable by every one;
[ ] Convert it into a PWA to be able to use it offline.

Once those features will be there, this project will be considered as
finished.